- Renamed file "___FILEBASENAME___Protocols.swift" to "___FILEBASENAME___ContractProtocols.swift" - Separated responsibility of interaction by dividing into two protocols 	1. "___FILEBASENAMEASIDENTIFIER___InteractorInputProtocol" i.e.Actual interaction 	2. "___FILEBASENAMEASIDENTIFIER___InteractorOutputProtocol" i.e. object of presenter inside the interactor, because interactor can only send the output to presenter - Removed init() method from presenter as assembling is the responsibility of the Router only during generating/creating/assembling module - Added some comments in "___FILEBASENAME___ContractProtocols.swift" so VIPER Noobs can easily understand the usage of each protocols - added get-set properties in ___FILEBASENAMEASIDENTIFIER___PresenterProtocol as this is the skeleton on presentor 	weak var view: ___FILEBASENAMEASIDENTIFIER___ViewProtocol? { get set } 	var interactor: ___FILEBASENAMEASIDENTIFIER___InteractorInputProtocol! { get set }     	var router: ___FILEBASENAMEASIDENTIFIER___WireframeProtocol! { get set } - Addition in "___FILEBASENAMEASIDENTIFIER___WireframeProtocol" 	- weak var viewController: UIViewController? { get set } 	- static func assembleModule() -> UIViewController - changes in "___FILEBASENAMEASIDENTIFIER___Interactor" 	- changed the type of presenter from  ___FILEBASENAMEASIDENTIFIER___PresenterProtocol to ___FILEBASENAMEASIDENTIFIER___InteractorOutputProtocol as interactor can only send the output to presenter - changes in ?(optional) to (!)unwrapped optional at few places
