+++ Module VIPER.xctemplate/TemplateInfo.plist
@@ -45,6 +45,8 @@
 			<string>static</string>
 		</dict>
 	</array>
+	<key>Template Impoved</key>
+	<string>Hitendra Hckr https://github.com/hitendradeveloper</string>
 	<key>Template Author</key>
 	<string>Juanpe Catalán - @JuanpeCMiOS</string>
 </dict>
