+++ Module VIPER.xctemplate/___FILEBASENAME___ContractProtocols.swift
@@ -1,0 +1,45 @@
+//
+//  ___FILENAME___
+//  ___PROJECTNAME___
+//
+//  Created ___FULLUSERNAME___ on ___DATE___.
+//  Copyright © ___YEAR___ ___ORGANIZATIONNAME___. All rights reserved.
+//
+//  Template generated by Juanpe Catalán @JuanpeCMiOS
+//  Template improved by Hitendra Hckr https://github.com/hitendradeveloper
+//
+
+import UIKit
+
+//MARK: View -
+protocol ___FILEBASENAMEASIDENTIFIER___ViewProtocol: class {
+    var presenter: ___FILEBASENAMEASIDENTIFIER___PresenterProtocol!  { get set }
+    //methods that receives data from Presenter
+}
+
+//MARK: Presenter -
+protocol ___FILEBASENAMEASIDENTIFIER___PresenterProtocol: class {
+    weak var view: ___FILEBASENAMEASIDENTIFIER___ViewProtocol? { get set }
+    var interactor: ___FILEBASENAMEASIDENTIFIER___InteractorInputProtocol! { get set }
+    var router: ___FILEBASENAMEASIDENTIFIER___WireframeProtocol! { get set }
+    
+    //methods to request data/userActions to Presenter
+}
+
+//MARK: Presenter - [Interactor-OUTPUT]
+protocol ___FILEBASENAMEASIDENTIFIER___InteractorOutputProtocol: class {
+    //methods that sends OUTPUT to Presenter
+}
+
+//MARK: Interactor - [Interactor-INPUT]
+protocol ___FILEBASENAMEASIDENTIFIER___InteractorInputProtocol: class {
+    var presenter: ___FILEBASENAMEASIDENTIFIER___InteractorOutputProtocol!  { get set }
+    //methods that receives request from Presenter to generate output
+}
+
+//MARK: Wireframe | Router -
+protocol ___FILEBASENAMEASIDENTIFIER___WireframeProtocol: class {
+    weak var viewController: UIViewController? { get set }
+    
+    static func assembleModule() -> UIViewController
+}
