+++ Module VIPER.xctemplate/___FILEBASENAME___Interactor.swift
@@ -6,11 +6,12 @@
 //  Copyright © ___YEAR___ ___ORGANIZATIONNAME___. All rights reserved.
 //
 //  Template generated by Juanpe Catalán @JuanpeCMiOS
+//  Template improved by Hitendra Hckr https://github.com/hitendradeveloper
 //
 
 import UIKit
 
-class ___FILEBASENAMEASIDENTIFIER___Interactor: ___FILEBASENAMEASIDENTIFIER___InteractorProtocol {
+class ___FILEBASENAMEASIDENTIFIER___Interactor: ___FILEBASENAMEASIDENTIFIER___InteractorInputProtocol {
 
-    weak var presenter: ___FILEBASENAMEASIDENTIFIER___PresenterProtocol?
+    weak var presenter: ___FILEBASENAMEASIDENTIFIER___InteractorOutputProtocol!
 }
