+++ Module VIPER.xctemplate/___FILEBASENAME___Presenter.swift
@@ -6,22 +6,14 @@
 //  Copyright © ___YEAR___ ___ORGANIZATIONNAME___. All rights reserved.
 //
 //  Template generated by Juanpe Catalán @JuanpeCMiOS
+//  Template improved by Hitendra Hckr https://github.com/hitendradeveloper
 //
 
 import UIKit
 
-class ___FILEBASENAMEASIDENTIFIER___Presenter: ___FILEBASENAMEASIDENTIFIER___PresenterProtocol {
+class ___FILEBASENAMEASIDENTIFIER___Presenter: ___FILEBASENAMEASIDENTIFIER___PresenterProtocol, ___FILEBASENAMEASIDENTIFIER___InteractorOutputProtocol {
 
-    weak private var view: ___FILEBASENAMEASIDENTIFIER___ViewProtocol?
-    private let interactor: ___FILEBASENAMEASIDENTIFIER___InteractorProtocol
-    private let router: ___FILEBASENAMEASIDENTIFIER___WireframeProtocol
-
-    init(interface: ___FILEBASENAMEASIDENTIFIER___ViewProtocol, interactor: ___FILEBASENAMEASIDENTIFIER___InteractorProtocol, router: ___FILEBASENAMEASIDENTIFIER___WireframeProtocol) {
-        self.view = interface
-        self.interactor = interactor
-        self.router = router
-
-        self.interactor.presenter = self
-    }
-
+    weak var view: ___FILEBASENAMEASIDENTIFIER___ViewProtocol?
+    var interactor: ___FILEBASENAMEASIDENTIFIER___InteractorInputProtocol!
+    var router: ___FILEBASENAMEASIDENTIFIER___WireframeProtocol!
 }
