--- Module VIPER.xctemplate/___FILEBASENAME___Protocols.swift
@@ -1,32 +1,0 @@
-//
-//  ___FILENAME___
-//  ___PROJECTNAME___
-//
-//  Created ___FULLUSERNAME___ on ___DATE___.
-//  Copyright © ___YEAR___ ___ORGANIZATIONNAME___. All rights reserved.
-//
-//  Template generated by Juanpe Catalán @JuanpeCMiOS
-//
-
-import Foundation
-
-//MARK: Wireframe -
-protocol ___FILEBASENAMEASIDENTIFIER___WireframeProtocol: class {
-
-}
-//MARK: Presenter -
-protocol ___FILEBASENAMEASIDENTIFIER___PresenterProtocol: class {
-
-}
-
-//MARK: Interactor -
-protocol ___FILEBASENAMEASIDENTIFIER___InteractorProtocol: class {
-
-  var presenter: ___FILEBASENAMEASIDENTIFIER___PresenterProtocol?  { get set }
-}
-
-//MARK: View -
-protocol ___FILEBASENAMEASIDENTIFIER___ViewProtocol: class {
-
-  var presenter: ___FILEBASENAMEASIDENTIFIER___PresenterProtocol?  { get set }
-}
