+++ Module VIPER.xctemplate/___FILEBASENAME___Router.swift
@@ -6,6 +6,7 @@
 //  Copyright © ___YEAR___ ___ORGANIZATIONNAME___. All rights reserved.
 //
 //  Template generated by Juanpe Catalán @JuanpeCMiOS
+//  Template improved by Hitendra Hckr https://github.com/hitendradeveloper
 //
 
 import UIKit
@@ -14,12 +15,16 @@ class ___FILEBASENAMEASIDENTIFIER___Router: ___FILEBASENAMEASIDENTIFIER___Wirefr
     
     weak var viewController: UIViewController?
     
-    static func createModule() -> UIViewController {
+    static func assembleModule() -> UIViewController {
         // Change to get view from storyboard if not using progammatic UI
         let view = ___FILEBASENAMEASIDENTIFIER___ViewController(nibName: nil, bundle: nil)
         let interactor = ___FILEBASENAMEASIDENTIFIER___Interactor()
         let router = ___FILEBASENAMEASIDENTIFIER___Router()
-        let presenter = ___FILEBASENAMEASIDENTIFIER___Presenter(interface: view, interactor: interactor, router: router)
+        let presenter = ___FILEBASENAMEASIDENTIFIER___Presenter()
+        
+        presenter.view = interface
+        presenter.interactor = interactor
+        presenter.router = router
         
         view.presenter = presenter
         interactor.presenter = presenter
