+++ Module VIPER.xctemplate/___FILEBASENAME___ViewController.swift
@@ -6,13 +6,14 @@
 //  Copyright © ___YEAR___ ___ORGANIZATIONNAME___. All rights reserved.
 //
 //  Template generated by Juanpe Catalán @JuanpeCMiOS
+//  Template improved by Hitendra Hckr https://github.com/hitendradeveloper
 //
 
 import UIKit
 
 class ___FILEBASENAMEASIDENTIFIER___ViewController: UIViewController, ___FILEBASENAMEASIDENTIFIER___ViewProtocol {
 
-	var presenter: ___FILEBASENAMEASIDENTIFIER___PresenterProtocol?
+	var presenter: ___FILEBASENAMEASIDENTIFIER___PresenterProtocol!
 
 	override func viewDidLoad() {
         super.viewDidLoad()
